rename with different casing
