+++ Package.swift
@@ -21,7 +21,7 @@ let package = Package(
         .library(name: "BloomFilterWrapper", targets: ["BloomFilterWrapper"]),
         .library(name: "UserScript", targets: ["UserScript"]),
         .library(name: "Crashes", targets: ["Crashes"]),
-        .library(name: "cxxCrashHandler", targets: ["cxxCrashHandler"]),
+        .library(name: "CxxCrashHandler", targets: ["CxxCrashHandler"]),
         .library(name: "ContentBlocking", targets: ["ContentBlocking"]),
         .library(name: "PrivacyDashboard", targets: ["PrivacyDashboard"]),
         .library(name: "Configuration", targets: ["Configuration"]),
@@ -149,10 +149,10 @@ let package = Package(
             name: "Crashes",
             dependencies: [
                 "Common",
-                "cxxCrashHandler",
+                "CxxCrashHandler",
             ]),
         .target(
-            name: "cxxCrashHandler",
+            name: "CxxCrashHandler",
             dependencies: ["Common"]
         ),
         .target(
