+++ Sources/CxxCrashHandler/include/CxxCrashHandler.h
