+++ Sources/CxxCrashHandler/include/NSException+cxxHandler.h
