+++ Sources/CxxCrashHandler/include/TestException.h
